fix(video): do not list the same mode twice
Do not provide more than one entry in the settings for the same actual
mode, even if it was selected for different expected resolutions.

For example, do not list 640x480 twice for both 480p and 360p if the
device doesn't have a better 360p mode, and simply skip the 360p entry.