docs: remove mention of manual translation without Weblate
In the past year or so none of the translation pull requests that were
made directly without using Weblate were merged, since either they
didn't conform to the commit message format, or they conflicted with
translations pulled from Weblate.  This clearly shows that potential
translators should not be pointed out that it's possible to translate
without using Weblate, as it would only lead to wasting their time.