fix(groups): Correct color of labels in group call
Group calls are supposed to show the name of each member under their
avatars. The color of the text was previously fixed to white regardless
of the background ignoring the color of the background.

This fix ensures that the background color is not the same color as the
label text