fix(appimage): Include additional needed glib for RH based distros
Add old glib to work with our old gnutls on redhat based systems.

Generally we would want to avoid providing our old libs that the distro's newer
libs that we use depend on, which would include harfbuzz in this case for glib,
but including harfbuzz breaks libfreetype on newer distros like arch.

Other libs that we use from the system with dependencies that we provide include
are libfreetype, libXdmcp, and libICE, but that is unchanged here.

Fix #6339