chore(CI): Don't try to upload installer on debug builds
The installer is only built on release builds. We used to ignore failure to
upload prior to 9e449cbe6e1b528686a6d62c0525a154584da2aa. We stopped ignoring
because a471345388cfef98c9698ec12db8fe144331e941 tried to make our upload
targetted, but it only resolved the issue for the zip, not the installer.