feat(history): Remove group history on quit when selected
The same user may have sent a message both in the group and 1:1, so
generically remove aliases that have no associated messages, rather than
aliases that were used in the group.

Similarly, generically remove authors that have no associated aliases.

The group itself is a chat so will be removed through existing query.