perf(Settings): Make personal settings saving consistently async
The function is documented as async, but is only async when called from
a different thread. Save is called from within Settings, it looks like
only from public functions that would be called on other threads, but
make it async regardless of calling thread for consistency.

There is already a sync function that is used, so safety is not lost.