fix: Don't crash on quit during AV call.
The problem here was that during the destruction of ToxFriendCall's,
audio frames can still be sending. The fix here is very ugly: instead of
connecting the audioSource to the call, we directly connect it to
CoreAV. This is the same as we currently do for video source -> CoreAV.
It's unclear to me whether this is at all safe to do, so maybe some
weirdness happens to calls during shutdown, but at least no undefined
behaviour should happen during ToxFriendCall destruction now.