fix: Avoid occasional crash when changing video devices or closing qTox.
There was an AV race condition caused by clearing the frame buffer while
the VideoFrame is in flight towards toxav. The fix is heavy-handed in
that we're now just copying the frame every time, even though we almost
never have this problem (only when changing devices). So we pay the cost
all the time even though it's rarely needed. However, this is a serious
crashing bug (heap-use-after-free), so I'd rather have it fixed.