fix(Chat): Always show the date at the top of the chat.
The intended feature is to not show that date when we're in the current
date. The problem with that feature is that it doesn't work properly and
causes janky scrolling when moving over dates. This fixes that problem
by removing the feature. Maybe in the future we can revisit it.