fix(Startup): Allow Tox Address in the friend address settings.
We parse it as Tox public key, but allow nospam+checksum suffixes.

Also, don't crash when the Tox public key is the wrong size. It'll
probably do weird things afterwards, but there seem to be real users
with those broken (0-sized) keys in their settings. I'd rather not load
them, but the NDEBUG behaviour was to accept them, so we keep them like
this for now.

Fixes #562. Fixes #567.