fix: Call the correct QString::args() overload in Qt 6.9
Qt 6.9.0 redesigned the QString::args() overloads, making sure that
character-like types always output a character instead of a number. This
results in the compile failure, as there is no longer an overload for
QString::arg(const char32_t&, int, int, QChar). To get the old behavior
of printing the number and to fix the compile failure, we must cast the
character to an integral type, so that the integral overload is called.

See the commit message of:
https://github.com/qt/qtbase/commit/563ed822f867c6c3040956017d4ca7f3795f172c

Fixes TokTok/qTox#585